feat: initial public release — PashuRaksha ERP livestock management platform
Full-stack monorepo: FastAPI backend, Next.js admin, React Native mobile,
Vite collection/vet dashboards, mock services, Docker Compose stack.

Includes CI/CD, e2e/load/integration tests, seed data, and documentation.
Hardened .gitignore to exclude credentials, IDE config, and local state.

+++ pashu-erp/packages/api/app/schemas/income.py
@@ -1,19 +1,20 @@
 from datetime import date
+from decimal import Decimal
 
 from pydantic import BaseModel
 
 
 class IncomeSummary(BaseModel):
-    total_income: float
+    total_income: Decimal
     period: str  # "week" / "month" / "year"
     start_date: date
     end_date: date
-    by_product: dict[str, float]
+    by_product: dict[str, Decimal]
     transaction_count: int
 
 
 class IncomeBreakdown(BaseModel):
     product_type: str
-    amount: float
+    amount: Decimal
     percentage: float
     transaction_count: int
